made a basic auth

+++ src/controllers/task.controller.js
@@ -4,6 +4,9 @@ const prisma = require('../config/prisma');
 const getTasks = async (req, res) => {
   try {
     const tasks = await prisma.task.findMany({
+      where: {
+        userId: req.userId,
+      },
       orderBy: {createdAt: 'desc'},
     });
     res.json(tasks);
@@ -20,6 +23,7 @@ const createTask = async (req, res) => {
       data: {
         title: req.body.title,
         completed: req.body.completed || false,
+        userId: req.userId,
       },
     });
     res.status(201).json(task);
@@ -86,7 +90,18 @@ const patchTask = async (req, res) => {
 const deleteTask = async (req, res) => {
   try {
     const id = parseInt(req.params.id);
-    if (isNaN(id)) return res.status(400).json({error: 'Invalid ID'});
+    if (isNaN(id)) return res.status(400).json({ error: 'Invalid ID' });
+
+    const task = await prisma.task.findFirst({
+      where: {
+        id,
+        userId: req.userId
+      }
+    });
+
+    if (!task) {
+      return res.status(404).json({error: 'Task not found'})
+    }
 
     await prisma.task.delete({where: {id}});
     res.json({success: true, message: 'Задача удалена', deletedId: id});
